% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{abund_rect}
\alias{abund_rect}
\title{-----------------------------------------------------------------------------
Local species abundance distribution}
\usage{
abund_rect(x0, y0, xsize, ysize, comm)
}
\arguments{
\item{x0}{x-coordinate of lower left corner}

\item{y0}{x-coordinate of lower left corner}

\item{xsize}{size of the subplot in x-direction}

\item{ysize}{size of the subplot in y-direction}

\item{comm}{\code{\link{community}} object}
}
\value{
Integer vector with local species abundances
}
\description{
Get local abundance distribution in rectangle bounded by x0, y0, x0 + xsize,
y0 + ysize
}
