% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_ENS.R
\name{calc_ENS}
\alias{calc_ENS}
\title{Calculate ENS}
\usage{
calc_ENS(x, ENS = TRUE)
}
\arguments{
\item{x}{community matrix or vector}

\item{ENS}{boolian. Should ENS be returned? If \code{ENS=FALSE}, PIE will be returned.}
}
\description{
ENS is calculated from the unbiased PIE.
}
