% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{community}
\alias{community}
\title{Create spatial community object.}
\usage{
community(x, y, spec_id, xrange = c(0, 1), yrange = c(0, 1))
}
\arguments{
\item{x}{numeric - x coordinates}

\item{y}{numeric - y coordinates}

\item{spec_id}{character vector - species names or IDs. Can be integers, characters or factors}

\item{xrange}{numeric vector of length 2 - extent of the community in x-direction}

\item{yrange}{numeric vector of length 2 - extent of the community in y-direction}
}
\value{
community object which includes three items:
\enumerate{
   \item census - data.frame with three columns: X, Y and species names for each individual
   \item x_min_max - extent of the community in x-direction
   \item y_min_max - extent of the community in y-direction
}
}
\description{
Creates a spatial community object with a certain extent and with coordinates
and species identities of all individuals in the community.
}
\examples{
x <- runif(100)
y <- runif(100)
species_names <- rep(paste("species",1:10, sep = ""), each = 10)

com1 <- community(x,y, species_names)
plot(com1)
summary(com1)

}
