% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{community2ppp}
\alias{community2ppp}
\title{Converts an object of class \code{\link[mobsim]{community}} into an object of
class \code{\link[spatstat.geom]{ppp}}.}
\usage{
community2ppp(comm)
}
\arguments{
\item{comm}{Community object}
}
\description{
Converts an object of class \code{\link[mobsim]{community}} into an object of
class \code{\link[spatstat.geom]{ppp}}.
}
\details{
For each species, the function first draws a random number of
mother points in the given \code{n_motherpoints_range}. Then the function
places as many mother points as needed in the window given by \code{xrange}
and \code{yrange}.
}
\examples{
# Integrated between mobsim::sim_thomas_coords() and spatstat.explore::marktable()
simdat <- mobsim::sim_thomas_coords(abund_vec = 11L:100L)
simdat <- community2ppp(comm = simdat)
spatstat.explore::marktable(X = simdat, N = 10L)

}
\author{
Alban Sagouis
}
