% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{community_to_sad}
\alias{community_to_sad}
\title{Get species abundance distribution from community object}
\usage{
community_to_sad(com)
}
\arguments{
\item{com}{}
}
\description{
Get species abundance distribution from community object
}
