% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_motherpoints.R
\name{create_motherpoints}
\alias{create_motherpoints}
\alias{plot.mother_map}
\title{Creates a random number of mother points for a given number of species}
\usage{
create_motherpoints(
  s_local,
  n_motherpoints_range = 1:4,
  xrange = c(0, 1),
  yrange = c(0, 1),
  seed = NULL
)

\method{plot}{mother_map}(x, col = NULL, pch = NULL, ...)
}
\arguments{
\item{s_local}{(integer) Species number.}

\item{n_motherpoints_range}{(integer vector) Vector of integer values for the number of motherpoints a species can get.}

\item{xrange}{(numeric vector) Extent of the community in x-direction.}

\item{yrange}{(numeric vector) Extent of the community in y-direction.}

\item{seed}{(integer) Any integer passed to \code{\link[base]{set.seed}} for reproducibility.}

\item{x}{mother_map object}

\item{col}{Colour vector to mark species identities}

\item{pch}{Plotting character to mark species identities.
pch 16 is advised for large data sets.}

\item{...}{Other parameters to \code{\link[graphics:plot.default]{graphics::plot}}}
}
\value{
A named list of class \code{mother_map} with three elements:
\enumerate{
  \item   \code{n_mother_points} a vector of integers giving the number of mother points per species
  \item   \code{xmother} List of length equal to the number of species. Each list element
is a vector of x coordinates for every mother points. If one element is NA, the
the corresponding species is not clustered.
  \item   \code{ymother} List of length equal to the number of species. Each list element
is a vector of y coordinates for every mother points. If one element is NA, the
the corresponding species is not clustered.
}
}
\description{
Plot positions and species identities of all mother points created by
\code{\link{create_motherpoints}} or \code{\link{jitter_motherpoints}}.
}
\details{
For each species, the function first draws a random number of
mother points in the given \code{n_motherpoints_range}. Then the function
places as many mother points as needed in the window given by \code{xrange}
and \code{yrange}.
}
\examples{
# Default behaviour
mpcoords <- create_motherpoints(s_local = 10L)

# With more variability for the number of motherpoints
mpcoords <- create_motherpoints(s_local = 10L, n_motherpoints_range = 1L:100L)

# Allowing no clustering at all
mpcoords <- create_motherpoints(s_local = 10L, n_motherpoints_range = 0L:5L)

# Fixing the seed
mpcoords <- create_motherpoints(10L, seed = 42L)

# Integration with mobsim::sim_thomas_coords()
abund_vec <- 11L:100L # a community with 90 species
s_local <- length(abund_vec)
mpcoords <- create_motherpoints(s_local, seed = 42L)
simdat <- mobsim::sim_thomas_coords(abund_vec,
 xmother = mpcoords$xmother,	# list of vectors
 ymother = mpcoords$ymother	# list of vectors
)
# Integration with mobsim::sim_thomas_coords() without clustering
abund_vec <- 20L:25L
s_local <- length(abund_vec)
mpcoords <- create_motherpoints(s_local, n_motherpoints_range = 0L:1L, seed = 42L)
simdat <- mobsim::sim_thomas_coords(abund_vec,
 xmother = mpcoords$xmother,	# list of vectors
 ymother = mpcoords$ymother	# list of vectors
)
plot(simdat)

mpcoords <- create_motherpoints(10L)
plot(mpcoords)

}
\author{
Alban Sagouis
}
