% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{dist_decay}
\alias{dist_decay}
\title{Distance decay}
\usage{
dist_decay(comm, prop.A = 0.05, nsamples = 30, method = "bray",
  binary = F)
}
\arguments{
\item{comm}{\code{\link{community}} object}

\item{prop.A}{Subplot size as proportion of the total area}

\item{nsamples}{Number of randomly located subplots per subplot size}

\item{method}{Choise of (dis)similarity index. See \code{\link[vegan]{vegdist}}}

\item{binary}{Perform presence/absence standardization before analysis.}
}
\value{
Dataframe with distances between subplots and the respective similarity
indices
}
\description{
Estimate pairwise similarities of abundance distributions of subplots as
function of distance
}
\examples{
sim_com1 <- Sim.Thomas.Community(100, 10000)
dd1 <- dist_decay(sim_com1)
plot(similarity ~ distance, data = dd1)
dd_loess <- loess(similarity ~ distance, data = dd1)
new_dist <- data.frame(distance = seq(0.01,1,0.01))
pred_dd <- predict(dd_loess, newdata = new_dist$distance)
lines(new_dist$distance, pred_dd, lwd=2, col = "red")


}

