% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{div_rand_rect}
\alias{div_rand_rect}
\title{-----------------------------------------------------------------------------
Distribution of local diversity indices}
\usage{
div_rand_rect(prop_area = 0.25, comm, n_rect = 100, exclude_zeros = F)
}
\arguments{
\item{prop_area}{Size of subplots as proportion of the total area}

\item{comm}{\code{\link{community}} object}

\item{n_rect}{Number of randomly located subplots}

\item{exclude_zeros}{logical - should subplots without individuals be excluded?}
}
\value{
Vector with mean and standard deviation of the following diversity
indices: (1) Number of species (2) Number of endemics (3) Shannon-diversity
(4) Simpson diversity
}
\description{
Get mean and sd of diversity indices in several equally sized subplots
of a community
}
\seealso{
\code{\link{div_rect}}
}

