% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{div_rect}
\alias{div_rect}
\title{Local diversity indices}
\usage{
div_rect(x0, y0, xsize, ysize, comm)
}
\arguments{
\item{x0}{x-coordinate of lower left corner}

\item{y0}{x-coordinate of lower left corner}

\item{xsize}{size of the subplot in x-direction}

\item{ysize}{size of the subplot in y-direction}

\item{comm}{\code{\link{community}} object}
}
\value{
Named vector with four values
\enumerate{
   \item n_species - the number of species
   \item n_endemics - the number of endemics
   \item shannon - Shannon diversity index defined as \eqn{H = - \sum p_{i} * log(p_[i])},
   where \eqn{p_i} is the relative abundance of species i:
   \item simpson - Simpson diversity index (= probabiliy of interspecific encounter PIE)
   defined as \eqn{D =  1 - \sum p_i^2}
}
}
\description{
Get diversity indices including species richness, no. of endemics,
Shannon and Simpson diversity for one rectangle subplot in the community
}
