% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{DivAR}
\alias{DivAR}
\title{Estimate diversity-area relationships}
\usage{
DivAR(community, prop.A = seq(0.1, 1, by = 0.1), nsamples = 100,
  xext = c(0, 1), yext = c(0, 1), exclude_zeros = F)
}
\arguments{
\item{community}{Dataframe with three columns: x, y, species identity}

\item{prop.A}{Proportion of area sampled}

\item{xext}{Minimum and maximum x-coordinates of the community}

\item{yext}{Minimum and maximum y-coordinates of the community}

\item{nrect}{Number of randomly located subplots}
}
\value{
Vector with mean and standard deviation of the following diversity
indices: (1) Number of species (2) Number of endemics (3) Shannon-diversity
(4) Simpson diversity
}
\description{
Estimate diversity-area relationships
}

