% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{diversity.rand.rect}
\alias{diversity.rand.rect}
\title{Get mean and sd of diversity indices in several equally sized subplots
of a community}
\usage{
diversity.rand.rect(prop.A = 0.25, community, nrect = 100, xext = c(0, 1),
  yext = c(0, 1))
}
\arguments{
\item{prop.A}{Proportion of the total area samples}

\item{community}{Dataframe with three columns: x, y, species identity}

\item{nrect}{Number of randomly located subplots}

\item{xext}{Minimum and maximum x-coordinates of the community}

\item{yext}{Minimum and maximum y-coordinates of the community}
}
\value{
Vector with mean and standard deviation of the following diversity
indices: (1) Number of species (2) Number of endemics (3) Shannon-diversity
(4) Simpson diversity
}
\description{
Get mean and sd of diversity indices in several equally sized subplots
of a community
}

