% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{diversity.rect}
\alias{diversity.rect}
\title{Get diversity indices for one rectangle subplot in the community}
\usage{
diversity.rect(x0, y0, xsize, ysize, community)
}
\arguments{
\item{x0}{x-coordinate of lower left corner}

\item{y0}{x-coordinate of lower left corner}

\item{xsize}{size of the subplot in x-direction}

\item{ysize}{size of the subplot in x-direction}

\item{community}{Dataframe with three columns: x, y, species identity}
}
\value{
Named vector with four values
\enumerate{
   \item nSpecies - the number of species
   \item nEndemics - the number of endemicsw
   \item shannon - Shannon diversity index defined as \eqn{H = - \sum p_{i} * log(p_[i])},
   where \eqn{p_i} is the relative abundance of species i:
   \item simpson - Simpson diversity index (= probabiliy of interspecific encounter PIE)
   defined as \eqn{D =  1 - \sum p_i^2}
}
}
\description{
Get diversity indices for one rectangle subplot in the community
}

