% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_species.R
\name{drift_x_species}
\alias{drift_x_species}
\alias{drift_y_species}
\title{Adds the same fixed movement to all individuals of a species on x}
\usage{
drift_x_species(comm, drift = NULL)

drift_y_species(comm, drift = NULL)
}
\arguments{
\item{comm}{(community) List of class \code{community}.}

\item{drift}{(numeric) A single value added to the x or y coordinates.}
}
\value{
A named list of class \code{community} as described in
\code{\link[mobsim]{community}}
}
\description{
Adds the same fixed movement to all individuals of a species on x

Adds the same fixed movement to all individuals of a species on y
}
\details{
The function takes a \code{community} object and adds a fixed
drift value to either x or y coordinate per species. In a 1 by 1 landscape, a
drift of 0.5 means moving over half the map.
}
\examples{
# Default behaviour
simdat <- mobsim::sim_thomas_community(s_pool = 5L, n_sim = 100L,
                                       mother_points = 1L)
simdatD <- simdat |>
             drift_x_species(0.1) |>
             drift_y_species(-0.2)
par(mfrow = c(1L, 2L))
plot(simdat)
plot(simdatD)

}
\seealso{
drift_motherpoints
}
\author{
Alban Sagouis
}
