% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_species.R
\name{jitter_species}
\alias{jitter_species}
\title{Adds the same random movement to all individuals of a species}
\usage{
jitter_species(comm, sd = 0.01, seed = NULL)
}
\arguments{
\item{comm}{(community) List of class \code{community}.}

\item{sd}{(numeric) Standard deviation for the \code{\link[stats]{rnorm}}.}

\item{seed}{(integer) Any integer passed to \code{\link[base]{set.seed}} for reproducibility.}
}
\value{
A named list of class \code{community} as described in
\code{\link[mobsim]{community}}
}
\description{
Adds the same random movement to all individuals of a species
}
\details{
The function takes a \code{community} object and adds a random
jitter value to each coordinate per species. The jitter follows a normal
probability distribution of \code{mean} = 0 and \code{sd} specified by the
used with argument \code{sd}.
}
\examples{
# Default behaviour
simdat <- mobsim::sim_thomas_community(s_pool = 5L, n_sim = 100L,
                                       mother_points = 1L)
simdatJ <- jitter_species(simdat)
par(mfrow = c(1L, 2L))
plot(simdat)
plot(simdatJ)

# With more variability for the movement of individuals
simdatJ <- jitter_species(simdat, sd = 2)

# Fixing the seed
simdatJ <- jitter_species(simdat, seed = 42L)

}
\seealso{
jitter_motherpoints
}
\author{
Alban Sagouis
}
