% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobsim_package.R
\docType{package}
\name{mobsim-package}
\alias{mobsim}
\alias{mobsim-package}
\title{mobsim: A package for spatial analysis of scale-dependent biodiversity changes.}
\description{
The package includes functions to simulate species distributions in space
as well as for the analysis of spatially-explicit data, where each individual
is described by its xy-coordinates and a species identity label.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MoBiodiv/mobsim}
  \item Report bugs at \url{https://github.com/MoBiodiv/mobsim/issues}
}

}
\author{
\strong{Maintainer}: Felix May \email{felix.may@fu-berlin.de} (\href{https://orcid.org/0000-0002-1106-8188}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Alban Sagouis \email{alban.sagouis@idiv.de} (\href{https://orcid.org/0000-0002-3827-1063}{ORCID})
}

Other contributors:
\itemize{
  \item German Centre for Integrative Biodiversity Research (iDiv) Halle-Jena-Leipzig (01jty7g66) [funder]
  \item Freie Universität Berlin (046ak2485) [funder]
}

}
\keyword{internal}
