% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobsim.R
\docType{package}
\name{mobsim}
\alias{mobsim-package}
\alias{mobsim}
\title{mobsim: A package for spatial analysis of scale-dependent biodiversity changes.}
\description{
The package includes functions to simulate species distributions in space
as well as for the analysis of spatially-explicit data, where each individual
is described by its xy-coordinates and a species identity label.
}
\section{Functions to simulate species abundances and distributions}{


\code{\link{sim_sad}}

\code{\link{sim_poisson_coords}}

\code{\link{sim_thomas_coords}}

\code{\link{sim_poisson_community}}

\code{\link{sim_thomas_community}}

\code{\link{rThomas_rcpp}}
}

\section{Functions to analyse species abundances and distributions}{


\code{\link{rare_curve}}

\code{\link{spec_sample_curve}}

\code{\link{divar}}

\code{\link{dist_decay}}

\code{\link{dist_decay_quadrats}}

\code{\link{sample_quadrats}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MoBiodiv/mobsim}
  \item Report bugs at \url{https://github.com/MoBiodiv/mobsim}
}

}
\author{
Felix May
}
