% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBspatial.R
\docType{package}
\name{MoBspatial}
\alias{MoBspatial}
\alias{MoBspatial-package}
\title{MoBspatial: A package for spatial analysis of scale-dependent biodiversity changes.}
\description{
The package includes functions to simulated species distributions in space
as well as for the analysis of spatially-explicit data, where each individual
is described by its xy-coordinates as well as a species label.
}
\section{Functions to simulate species abundances and distributions}{

SAD.lognorm()

Sim.Poisson.Coords()

Sim.Thomas.Coords()

Sim.Poisson.Community()

Sim.Thomas.Community()
}

\section{Functions to analyse species abundances and distributions}{

SAC.coleman()

sSAC.all.rcpp()

DivAR()

distance.decay()
}
\author{
Felix May
}

