% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{plot.community}
\alias{plot.community}
\title{Plot spatial community object}
\usage{
\method{plot}{community}(com1, col = NA, pch = NA, patterns = c("points",
  "sad", "rare_accum", "divar", "dist_decay"), ...)
}
\description{
Plot spatial community object
}

