% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{plot.dist_decay}
\alias{plot.dist_decay}
\title{Plot distance decay object}
\usage{
\method{plot}{dist_decay}(dist_decay)
}
\arguments{
\item{dist_decay}{}
}
\description{
Plot distance decay object
}

