% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{plot.sad}
\alias{plot.sad}
\title{Plot species abundance distribution}
\usage{
\method{plot}{sad}(abund, method = c("octave", "rank"))
}
\arguments{
\item{abund}{}

\item{method}{}
}
\description{
Plot species abundance distribution
}
