% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{plot.spec_sample_curve}
\alias{plot.spec_sample_curve}
\title{-----------------------------------------------------------------------------}
\usage{
\method{plot}{spec_sample_curve}(curve)
}
\description{
-----------------------------------------------------------------------------
}
