% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{rare_curve}
\alias{rare_curve}
\title{species rarefaction curves}
\usage{
rare_curve(abund.vec, method = "coleman")
}
\arguments{
\item{abund.vec}{integer vector with species abundance distribution}

\item{method}{available methods are \code{"coleman"} or \code{"hurlbert"}.
The species richness estimations of both methods quickly converge for larger
numbers of individuals}
}
\value{
Numeric Vector with expected species richness in samples of 1, 2, 3 ... N individuals
}
\description{
Estimate expected species richness as a function of sample size
}
\examples{
sim_com1 <- SAD.lognorm(100, 10000)
rc1 <- rare_curve(sim_com1$abund)
rc2 <- rare_curve(sim_com1$abund, method = "hurlbert")

plot(rc1, type = "l", log = "xy", xlab = "Sample size",
     ylab = "Expected species richness")
lines(1:length(rc2), rc2, lty = 2, col = 2)

}
\references{
Coleman, B. D. et al. 1982. Randomness, area, and species richness. Ecology 63, 1121-1133

Hurlbert, S.H. 1971. The nonconcept of species diversity: a critique and alternative parameters.
Ecology 52, 577-586.
}

