% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{S.single.scale}
\alias{S.single.scale}
\title{Expected species richness in a random sample of fixed size}
\usage{
S.single.scale(n, abund.vec)
}
\arguments{
\item{n}{Integer of length 1 - sample size}

\item{abund.vec}{Integer vector - species abundance distribution of the community}
}
\value{
Expected number of species in n individuals
}
\description{
Expected species richness in a random sample of fixed size
}

