% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{SAC.coleman}
\alias{SAC.coleman}
\title{Non-spatial species accumulation curve (SAC) after Coleman}
\usage{
SAC.coleman(abund.vec)
}
\arguments{
\item{abund.vec}{Integer vector with species abundance distribution}
}
\value{
Numeric Vector with expected species numbers in samples of 1, 2, 3 ... N individuals
}
\description{
Non-spatial species accumulation curve (SAC) after Coleman
}
\references{
Coleman, B. D. et al. 1983. Randomness, area, and species richness. Ecology 63, 1121-1133
}

