% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{SAC}
\alias{SAC}
\title{Non-spatial species accumulation curve (SAC) after Coleman}
\usage{
SAC(abund.vec, method = "coleman")
}
\arguments{
\item{abund.vec}{Integer vector with species abundance distribution}
}
\value{
Numeric Vector with expected species numbers in samples of 1, 2, 3 ... N individuals
}
\description{
Non-spatial species accumulation curve (SAC) after Coleman
}
\references{
Coleman, B. D. et al. (1982). Randomness, area, and species richness. Ecology 63, 1121-1133
Hurlbert, S.H. (1971). The nonconcept of species diversity: a critique and alternative parameters. Ecology 52, 577–586.
}

