% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{SAD.lognorm}
\alias{SAD.lognorm}
\title{Simulate log-normal species abundance distributions}
\usage{
SAD.lognorm(S, N, cv = 1)
}
\arguments{
\item{S}{scalar integer - number of species}

\item{N}{scalar integer - number of individuals}

\item{cv}{scalar numeric - coefficient of variation ( = sd/mean) in abundance}
}
\value{
List with five named items:
\enumerate{
   \item abund - simulated abundance vector
   \item mean.theor - theoretical mean of the distribution
   \item sd.theor - theoretical standard deviation
   \item mean.sim - simulated mean
   \item sd.sim - simulated standard deviation
}
}
\description{
Simulate log-normal abundance data with fixed number of individuals (N),
number of species (S) and fixed coefficient of variation in abundance (cv)
}
\details{
The function precisely controls for the number of species and
the number of individuals in the simulated community. These constraints
can result in biases from the theoretical parameters. Therefore the simulated
distribution parameters are provided as model output.
}
\examples{
SAD.lognorm(S = 100, N = 10000, cv = 1)

## Create preston SAD plot with log2 abundance classes

require(untb)

abund.vec <- SAD.lognorm(S = 100, N = 10000, cv = 1)$abund
names(abund.vec) <- paste("spec", 1:length(abund.vec),sep="")
sad1 <- preston(census(abund.vec))
barplot(height = as.numeric(sad1), names.arg = names(sad1),
        xlab = "Abundance class", ylab ="No. of species")
}
\author{
Felix May
}

