% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_quadrats.R
\name{sample_quadrats}
\alias{sample_quadrats}
\title{Plot-based samples from a spatially-explicit census}
\usage{
sample_quadrats(comm, n_quadrats = 10, quadrat_area = 0.01, plot = T,
  method = "random", avoid_overlap = F, x0 = 0, y0 = 0, delta_x = 0.1,
  delta_y = 0.1)
}
\arguments{
\item{comm}{Community object from which the samples are generated}

\item{n_quadrats}{Number of sampling quadrats}

\item{quadrat_area}{Area of the sampling quadrats}

\item{plot}{Logical: should the sampling design be plotted?}

\item{method}{Available methods are \code{"random", "transect", "grid"}}

\item{avoid_overlap}{Logical: for the random sampling try to generate a design
without overlap of quadrats}

\item{x0, }{y0 Lower left corner of the first quadrat in transect and grid sampling}

\item{delta_x}{Distance between consecutive quadrats in transect and grid sampling
in x-direction (the distance between the left sides is measured)}

\item{delta_y}{Distance between consecutive quadrats in transect and grid sampling
in y-direction (the distance between the lower sides is measured)}
}
\value{
A matrix with species abundances where the rows represents sampling
quadrats and the columns represent species
}
\description{
This function allows to sample quadratic subplots from a spatially-explicit
community. The output format is a sites x species abundance table which is
the classical data format used in community ecology. The table generated
can be for instance be further analysed with the package \code{\link[vegan]}.
}
\examples{
library(vegan)
sim_com1 <- Sim.Poisson.Community(100, 10000)
comm_mat1 <- sample_quadrats(sim_com1, n_quadrats = 100,
quadrat_area = 0.002, method = "grid")
renyi(comm_mat1, hill = T)

}
