% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_quadrats.R
\name{sampling_one_quadrat}
\alias{sampling_one_quadrat}
\title{Creates one square quadrat randomly located in the landscape}
\usage{
sampling_one_quadrat(xmin, xmax, ymin, ymax, seed = NULL)
}
\arguments{
\item{xmin}{(numeric) minimum possible value on the x axis a quadrat can cover.}

\item{xmax}{(numeric) maximum possible value on the x axis a quadrat can cover.}

\item{ymin}{(numeric) minimum possible value on the y axis a quadrat can cover.}

\item{ymax}{(numeric) maximum possible value on the y axis a quadrat can cover.}

\item{seed}{Integer. Any integer passed to \code{set.seed} for reproducibility.}
}
\value{
a data.frame with 2 columns x and y giving  the coordinates of the
lower left corner of the square quadrat.
}
\description{
Creates one square quadrat randomly located in the landscape
}
