% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_quadrats.R
\name{sampling_transects}
\alias{sampling_transects}
\title{Creates square quadrats aligned along a transect}
\usage{
sampling_transects(
  n_quadrats,
  xmin,
  xmax,
  ymin,
  ymax,
  x0,
  y0,
  delta_x,
  delta_y,
  quadrat_size,
  seed = NULL
)
}
\arguments{
\item{n_quadrats}{Number of sampling quadrats}

\item{xmin}{(numeric) minimum possible value on the x axis a quadrat can cover.}

\item{xmax}{(numeric) maximum possible value on the x axis a quadrat can cover.}

\item{ymin}{(numeric) minimum possible value on the y axis a quadrat can cover.}

\item{ymax}{(numeric) maximum possible value on the y axis a quadrat can cover.}

\item{x0}{Lower left corner of the first quadrat in transect and grid sampling}

\item{y0}{Lower left corner of the first quadrat in transect and grid sampling}

\item{delta_x}{Distance between consecutive quadrats in transect and grid sampling
in x-direction (the distance between the left sides is measured)}

\item{delta_y}{Distance between consecutive quadrats in transect and grid sampling
in y-direction (the distance between the lower sides is measured)}

\item{quadrat_size}{(numeric) width of the quadrats.}

\item{seed}{Integer. Any integer passed to \code{set.seed} for reproducibility.}
}
\value{
a data.frame with 2 columns x and y giving  the coordinates of the
lower left corner of the square quadrats.
}
\description{
Creates square quadrats aligned along a transect
}
