% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{Sim.Poisson.Community}
\alias{Sim.Poisson.Community}
\title{Simulate community with log-normal SAD and random spatial positions}
\usage{
Sim.Poisson.Community(S, N, cv.abund = 1, xmax = 1, ymax = 1, fix.S = F)
}
\arguments{
\item{S}{scalar integer - number of species}

\item{N}{scalar integer - number of individuals}

\item{xmax}{scalar numeric - size of the simulated area in x-direction}

\item{ymax}{scalar numeric - size of the simulated area in y-direction}

\item{cv}{scalar numeric - coefficient of variation ( = sd/mean) in abundance}
}
\value{
A dataframe with three columns
\enumerate{
   \item X : x-coordinate of individual
   \item Y : y-coordinate of individual
   \item SpecID : factor with species identity label (from 1 to number of species)
}
}
\description{
Simulate community with log-normal SAD and random spatial positions
}
\examples{
community1 <- Sim.Poisson.Community(S = 20, N = 500, cv = 1)

spec.cols <- rainbow(length(levels(community1$SpecID)))
plot(Y ~ X, data =community1, col = spec.cols[community1$SpecID], pch = 19)

}
\author{
Felix May
}

