% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{Sim.Poisson.Coords}
\alias{Sim.Poisson.Coords}
\title{Random spatial coordinates}
\usage{
Sim.Poisson.Coords(abund.vec, xrange = c(0, 1), yrange = c(0, 1))
}
\arguments{
\item{abund.vec}{integer vector with species abundances}

\item{xrange}{numeric vector of length 2 - extent of the community in x-direction}

\item{yrange}{numeric vector of length 2 - extent of the community in y-direction}
}
\value{
A community object as defined by \code{\link{community}}.
}
\description{
Add random spatial positions to a species abundance distribution.
}
\examples{
abund <- SAD.lognorm(S.pool = 100, N.local = 1000)$abund
sim.com1 <- Sim.Poisson.Coords(abund)
plot(sim.com1)
summary(sim.com1)
}
\author{
Felix May
}

