% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{Sim.Thomas.Community}
\alias{Sim.Thomas.Community}
\title{Simulate community with clumped spatial positions.}
\usage{
Sim.Thomas.Community(S.pool, N.local, cv.abund = 1, fixS.local = F,
  sigma = 0.02, cluster.points = NA, mother.points = NA, xrange = c(0,
  1), yrange = c(0, 1))
}
\arguments{
\item{S.pool}{integer - number of species in the pool}

\item{N.local}{integer - number of individuals in the local community}

\item{cv.abund}{numeric - coefficient of variation ( = sd/mean) of relative
abundances. The higher \code{cv.abund}, the lower the evenness of the
simulated community. This means with increasing \code{cv.abund} there are more
rare and more dominant species.}

\item{fixS.local}{logical - should the simulation constrain the number of
species in the local community? This can result in deviations from mean and
sd of local abundances from the theoretical distributions}

\item{sigma}{numeric vector of length = 1 or length = 2. Mean displacement
(along each coordinate axes) of a point from its mother point (=cluster centre).}

\item{cluster.points}{numeric - mean number of points per cluster.}

\item{mother.points}{numeric - number of mother points (= cluster centres).}

\item{xrange}{numeric vector of length 2 - extent of the community in x-direction}

\item{yrange}{numeric vector of length 2 - extent of the community in y-direction}
}
\value{
A community object as defined by \code{\link{community}}.
}
\description{
This function simulates a community with a certain abundance distribution and
and clumped, i.e. aggregated, spatial coordinates. This function consecutively calls
\code{\link{SAD.lognorm}} and \code{\link{Sim.Thomas.Coords}}
}
\examples{
com1 <- Sim.Thomas.Community(S.pool = 20, N.local = 500, cv.abund = 1,
                             sigma = 0.01)
plot(com1)

}
\author{
Felix May
}
\seealso{
\code{SAD.lognorm}; \code{Sim.Thomas.Coords}
}

