% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_ENS.R
\name{sim_ENS}
\alias{sim_ENS}
\title{Simulate lognorm community with prescribed ENS}
\usage{
sim_ENS(ENS_target, S, mean_abu = 10, n_sim = NULL, sims = 100)
}
\arguments{
\item{ENS_target}{positiv numeric. Desired ENS value of the simulated community.}

\item{S}{integer. Number of species to be simulated.}

\item{mean_abu}{mean species abundance. Will be used to calculate total number of individuals \code{n_sim}.}

\item{n_sim}{total number of individuals. If \code{n_sim=NULL}, it will be calculated as \code{mean_abu*S}}

\item{sims}{number of repetitions used during the optimisation process (using different random seeds, respectively).}
}
\description{
Wrapper for sim_sad optimizing parameters of lnorm
distribution to create community with desired ENS.
The unbiased version of ENS is used here.
}
