% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{sim_sad}
\alias{sim_sad}
\title{Simulate log-normal species abundance distributions}
\usage{
sim_sad(s_pool, n_sim, cv_abund = 1, fix_s_sim = FALSE)
}
\arguments{
\item{s_pool}{Number of species in the pool (integer)}

\item{n_sim}{Number of individuals in the simulated community (integer)}

\item{cv_abund}{Coefficient of variation ( = sd/mean) of relative
abundances. The higher \code{cv_abund}, the lower the evenness of the
simulated community. This means with increasing \code{cv_abund} there are more
rare and more dominant species (numeric)}

\item{fix_s_sim}{Should the simulation constrain the number of
species in the simulated local community? (logical)}
}
\value{
Object of class \code{sad}, which contains a named integer vector
 with species abundances
}
\description{
Simulate abundance data from a log-normal distributions with defined number
of species in the pool \code{s_pool}, defined number of individuals
\code{n_sim}, and defined coefficient of variation of relative abundances
\code{cv_abund}.
}
\details{
When \code{fix_s_sim = FALSE} the species number in the local
  community might deviate from \code{s_pool} due to stochastic sampling. When
  \code{fix_s_sim = TRUE} the local number of species will equal
  \code{s_pool}, but this constraint can result in systematic biases from the
  theoretical distribution parameters defined by \code{cv_abund}.
}
\examples{
abund1 <- sim_sad(s_pool = 100, n_sim = 10000, cv_abund = 1)
plot(abund1, method = "octave")
plot(abund1, method = "rank")

}
\author{
Felix May
}
