% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{spec_sample}
\alias{spec_sample}
\title{-----------------------------------------------------------------------------
Sample species richness}
\usage{
spec_sample(abund_vec, n)
}
\arguments{
\item{abund_vec}{integer vector - species abundance distribution of the community}

\item{n}{single integer - sample size in number of individuals}
}
\value{
expected number of species in n individuals
}
\description{
Expected species richness in a random sample of fixed size
}
\details{
The expected number of species is calculated after Coleman 1982.
}
\references{
Hurlbert, S.H. 1971. The nonconcept of species diversity: a critique and alternative parameters.
Ecology 52, 577-586.
}

