% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{spec_sample_curve}
\alias{spec_sample_curve}
\title{-----------------------------------------------------------------------------
Spatially-explicit species accumulation curve (SAC)}
\usage{
spec_sample_curve(comm, method = c("accumulation", "rarefaction"))
}
\arguments{
\item{comm}{\code{\link{community}} object}
}
\value{
integer vector with expected species number when n neighbouring individuals
         are sampled
}
\description{
The SAC is similar to the rarefaction curve, but samples of defined size
always contain nearest-neighbour individuals. Accordingly the SAC is
influenced by intraspecific aggregation, while the rarefaction curve is only
driven by the species abundance distribution.
}
\examples{
sim_com1 <- Sim.Thomas.Community(100, 1000)
sac1 <- SAC(sim_com1)
rare_curve1 <- rare_curve(table(sim_com1$census$Species))

plot(sac1, type = "l", col = 2, xlab = "Sample size",
      ylab = "Expected species richness")
lines(1:length(rare_curve1), rare_curve1, col = 1)
legend("bottomrigh",c("Rarefaction curve","Species accumulation curve"),
       col = 1:2, lwd = 2)#

}
