% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{sSAC.all.rcpp}
\alias{sSAC.all.rcpp}
\title{Spatially-explicit species accumulation curve (sSAC)}
\usage{
sSAC.all.rcpp(data)
}
\arguments{
\item{data}{Data.frame or matrix with three columns.
The first two columns need to include the x,y coordinates
of individuals and the third column the species identity label.}
}
\value{
integer vector with expected species number when n neighbouring individuals
         are sampled
}
\description{
The sSAC is calculated starting from all individuals in the local community
and then averaged over all focal individuals
}

