% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{summary.sad}
\alias{summary.sad}
\title{Print summary of species abundance distribution object}
\usage{
\method{summary}{sad}(object, ...)
}
\arguments{
\item{object}{Community object of class \code{sad}}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
Print summary of species abundance distribution object
}
\seealso{
\code{\link{sim_sad}}
}
