% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Community.R
\name{theme_biodiviz}
\alias{theme_biodiviz}
\title{Plotting basic sampling relationships of simulated community}
\usage{
theme_biodiviz(base_size = 12, base_family = "")
}
\arguments{
\item{print}{- logical. If TRUE plot is plotted directly, otherwise is returned with an object to be plotted later, default is TRUE}

\item{col}{- color to be used for plotting. Default is 'black'}

\item{lty}{- line type to be used for plotting. Default is 'solid'}

\item{size}{- size of lines. Default is 1.}

\item{add.ribbon}{- logical. If TRUE plots a 95% confidence interval around the species-area curve. Default is FALSE.}

\item{plot.to.add}{- output of ggplot}

\item{plot.type}{- character. Type of plot, either "SAD", "SAC", or "SAR".}
}
\description{
Plotting basic sampling relationships of simulated community
}
\examples{
sim.com <- Sim.Thomas.Community(S = 10, N = 100, sigma=0.02, cv = 1)
spatial.plot(sim.com)
SAD.plot(sim.com)
SAC.plot(sim.com)
SAR.plot(sim.com)

sim.com1 <- Sim.Thomas.Community(S = 10, N = 1000, sigma=0.02, cv = 1)
sim.com2 <- Sim.Thomas.Community(S = 100, N = 1000, sigma=0.02, cv = 1)
SAC.plot1 <- SAC.plot(sim.com1, print=F)
SAC.plot2 <- add.to.plot(sim.com2, plot.to.add=SAC.plot1, plot.type="SAC", col="red",print=T)
}

