% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{torusify}
\alias{torusify}
\alias{torusify.mother_map}
\alias{torusify.community}
\alias{torusify.data.frame}
\alias{torusify.numeric}
\title{Rounds coordinates around like on a torus}
\usage{
torusify(mpcoords, ...)

\method{torusify}{mother_map}(mpcoords, ...)

\method{torusify}{community}(mpcoords, ...)

\method{torusify}{data.frame}(mpcoords, x_min_max, y_min_max, ...)

\method{torusify}{numeric}(mpcoords, range, ...)
}
\arguments{
\item{mpcoords}{A vector of class \code{numeric}.}

\item{...}{Kept for future uses.}

\item{x_min_max}{vector of 2 values describing the window on the x axis}

\item{y_min_max}{vector of 2 values describing the window on the y axis}

\item{range}{vector of 2 values describing the window.}
}
\value{
Object returned depends on the object given. For an object of class
\code{mother_map} as given by \code{\link{create_motherpoints}}, see
\code{\link{torusify.mother_map}}. For a \code{data.frame}, see
\code{\link{torusify.data.frame}} and for a \code{numeric}  vector, see
\code{\link{torusify.numeric}}.

An object of class \code{mother_map} of the same dimensions as the
one provided to the function.

A list of class \code{community} of same dimension as
\code{mpcoords}.

A \code{data.frame} of same dimension as \code{mpcoords}.

A \code{numeric} vector of same length as \code{mpcoords}.
}
\description{
Rounds coordinates around like on a torus

Rounds coordinates around like on a torus

Rounds coordinates around like on a torus

Rounds coordinates around like on a torus

Rounds coordinates around like on a torus
}
\details{
\code{torusify()} is a generic function that calls functions adapted
to the class of argument \code{mapcoords}. Classes \code{data.frame},
 \code{numeric} and \code{mother_map} are accepted at the moment.

\code{torusify()} takes coordinates and makes sure that all
coordinates actually inside the window specified
in \code{x_min_max} and \code{y_min_max}.

The function takes a \code{mother_map} object and makes sure that
all coordinates that were just jittered are actually inside the window
specified in the \code{x_min_max} and \code{y_min_max} elements of the
\code{mother_map} object.

The function takes a \code{community} and makes sure that all
individual coordinates are actually inside the window specified by the
\code{x_min_max} and \code{y_min_max} elements of the \code{community}
object.

The function takes a \code{data.frame} and makes sure that all
coordinates are actually inside the window specified
by the \code{x_min_max} and \code{y_min_max} arguments.

The function takes a \code{numeric} vector and makes sure that all
values are actually inside the window specified
by the \code{range} argument.
}
\examples{
# Default behaviour
mpcoords <- create_motherpoints(s_local = 10L)
mpcoordsJ <- jitter_motherpoints(mpcoords)
mcoordsT <- torusify(mpcoordsJ)
# Default behaviour
mpcoords <- create_motherpoints(s_local = 10L)
mpcoordsJ <- jitter_motherpoints(mpcoords)
mcoordsT <- torusify(mpcoordsJ)

simdat <- mobsim::sim_thomas_community(10L, 100L, mother_points = 1)
simdatJ <- jitter_species(simdat, sd = 1)
simdatJ <- torusify(simdatJ)

x_min_max <- c(0, 1)
y_min_max <- c(0, 1)
mpcoords <- data.frame(xmother = stats::runif(100L),
                       ymother = stats::runif(100L))
mcoordsT <- torusify(mpcoords, x_min_max, y_min_max)

mpcoords <- stats::runif(10L, min = -1, max = 2)
torusify(mpcoords, range = c(0, 1))

}
\author{
Alban Sagouis
}
