% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{community_to_sad}
\alias{community_to_sad}
\title{Get species abundance distribution from community object}
\usage{
community_to_sad(comm)
}
\arguments{
\item{comm}{Community object}
}
\value{
Object of class \code{sad}, which contains a named integer vector
 with species abundances
}
\description{
Get species abundance distribution from community object
}
\examples{
sim1 <- sim_poisson_community(s_pool = 200, n_sim = 20000, sad_type = "lnorm",
                              sad_coef = list("cv_abund" = 2))
sad1 <- community_to_sad(sim1)
plot(sad1, method = "rank")
plot(sad1, method = "octave")

}
