% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBspatial.R
\docType{package}
\name{MoBspatial}
\alias{MoBspatial}
\alias{MoBspatial-package}
\title{MoBspatial: A package for spatial analysis of scale-dependent biodiversity changes.}
\description{
The package includes functions to simulate species distributions in space
as well as for the analysis of spatially-explicit data, where each individual
is described by its xy-coordinates and a species identity label.
}
\section{Functions to simulate species abundances and distributions}{


\code{\link{SAD.lognorm}}

\code{\link{Sim.Poisson.Coords}}

\code{\link{Sim.Thomas.Coords}}

\code{\link{Sim.Poisson.Community}}

\code{\link{Sim.Thomas.Community}}
}

\section{Functions to analyse species abundances and distributions}{


\code{\link{rare_curve}}

\code{\link{SAC}}

DivAR()

distance.decay()
}
\author{
Felix May
}

