% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{plot.community}
\alias{plot.community}
\title{Plot spatial community object}
\usage{
\method{plot}{community}(x, ..., col = NULL, pch = NULL)
}
\arguments{
\item{x}{Community object}

\item{...}{Other parameters to \link[graphics]{plot}}

\item{col}{Colour vector to mark species identities}

\item{pch}{Plotting character to mark species identities}
}
\description{
Plot positions and species identities of all individuals in a community object.
}
\examples{
sim1 <- sim_thomas_community(30, 500)
plot(sim1)

}
