% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{plot.dist_decay}
\alias{plot.dist_decay}
\title{Plot distance decay of similarity}
\usage{
\method{plot}{dist_decay}(x, ...)
}
\arguments{
\item{x}{Dataframe generated by \code{\link{dist_decay}}}

\item{...}{Additional graphical parameters used in \code{\link[graphics]{plot}}.}
}
\description{
Plot distance decay of similarity
}
\details{
The function plots the similarity indices between all pairs of
subplots as function of distance. To indicate the relationship a
\code{\link[stats]{loess}} smoother is added to the plot.
}
\examples{
sim_com1 <- sim_thomas_community(100, 10000)
dd1 <- dist_decay(sim_com1)
plot(dd1)

}
