% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Area_Relationships.R
\name{plot.divar}
\alias{plot.divar}
\title{-----------------------------------------------------------------------------
Plot diversity-area relationships}
\usage{
\method{plot}{divar}(x, ...)
}
\arguments{
\item{x}{Dataframe generated function \code{\link{divar}}.}

\item{...}{Additional graphical parameters used in \code{\link[graphics]{plot}}.}
}
\description{
-----------------------------------------------------------------------------
Plot diversity-area relationships
}
