% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_spatial.R
\name{plot.spec_sample_curve}
\alias{plot.spec_sample_curve}
\title{Plot species sampling curves}
\usage{
\method{plot}{spec_sample_curve}(x, ...)
}
\arguments{
\item{x}{Species sampling curve generated by \code{\link{spec_sample_curve}}}

\item{...}{Additional graphical parameters used in \code{\link[graphics]{plot}}.}
}
\description{
Plot species sampling curves
}
\examples{
sim_com1 <- sim_thomas_community(s_pool = 100, n_sim = 1000)
sac1 <- spec_sample_curve(sim_com1, method = c("rare","acc"))
plot(sac1)

}
