% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_quadrats.R
\name{sampling_random_overlap}
\alias{sampling_random_overlap}
\title{Creates square quadrats randomly distributed that may overlap each other}
\usage{
sampling_random_overlap(n_quadrats, xmin, xmax, ymin, ymax, quadrat_size)
}
\arguments{
\item{n_quadrats}{Number of sampling quadrats}

\item{xmin}{(numeric) minimum possible value on the x axis a quadrat can cover.}

\item{xmax}{(numeric) maximum possible value on the x axis a quadrat can cover.}

\item{ymin}{(numeric) minimum possible value on the y axis a quadrat can cover.}

\item{ymax}{(numeric) maximum possible value on the y axis a quadrat can cover.}

\item{quadrat_size}{(numeric) width of a square quadrat}
}
\value{
a data.frame with 2 columns x and y giving  the coordinates of the
lower left corner of the square quadrats.
}
\description{
Creates square quadrats randomly distributed that may overlap each other
}
