% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{Sim.Poisson.Community}
\alias{Sim.Poisson.Community}
\title{Simulate community with random spatial positions.}
\usage{
Sim.Poisson.Community(S.pool, N.local, cv.abund = 1, fixS.local = F,
  xrange = c(0, 1), yrange = c(0, 1))
}
\arguments{
\item{S.pool}{integer - number of species in the pool}

\item{N.local}{integer - number of individuals in the local community}

\item{cv.abund}{numeric - coefficient of variation ( = sd/mean) of relative
abundances. The higher \code{cv.abund}, the lower the evenness of the
simulated community. This means with increasing \code{cv.abund} there are more
rare and more dominant species.}

\item{fixS.local}{logical - should the simulation constrain the number of
species in the local community? This can result in deviations from mean and
sd of local abundances from the theoretical distributions}

\item{xrange}{numeric vector of length 2 - extent of the community in x-direction}

\item{yrange}{numeric vector of length 2 - extent of the community in y-direction}
}
\value{
A community object as defined by \code{\link{community}}.
}
\description{
This function simulates a community with a certain abundance distribution and
and random spatial coordinates. This function consecutively calls
\code{\link{SAD.lognorm}} and \code{\link{Sim.Poisson.Coords}}
}
\examples{
com1 <- Sim.Poisson.Community(S = 20, N = 500, cv = 1)
plot(com1)

}
\author{
Felix May
}

