% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{sim_poisson_community}
\alias{sim_poisson_community}
\title{Simulate community with random spatial positions.}
\usage{
sim_poisson_community(s_pool, n_sim, cv_abund = 1, fix_s_sim = F,
  xrange = c(0, 1), yrange = c(0, 1))
}
\arguments{
\item{s_pool}{Number of species in the pool (integer)}

\item{n_sim}{Number of individuals in the simulated community (integer)}

\item{cv_abund}{Coefficient of variation ( = sd/mean) of relative
abundances. The higher \code{cv_abund}, the lower the evenness of the
simulated community. This means with increasing \code{cv_abund} there are more
rare and more dominant species (numeric)}

\item{fix_s_sim}{Should the simulation constrain the number of
species in the simulated local community? (logical)}

\item{xrange}{Extent of the community in x-direction (numeric vector of length 2)}

\item{yrange}{Extent of the community in y-direction (numeric vector of length 2)}
}
\value{
A community object as defined by \code{\link{community}}.
}
\description{
This function simulates a community with a certain abundance distribution and
and random spatial coordinates. This function consecutively calls
\code{\link{sim_sad}} and \code{\link{sim_poisson_coords}}
}
\examples{
com1 <- sim_poisson_community(s_pool = 20, n_sim = 500, cv_abund = 1)
plot(com1)

}
\author{
Felix May
}
