% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{sim_sad}
\alias{sim_sad}
\title{Simulate species abundance distributions}
\usage{
sim_sad(s_pool, n_sim, sad_type = c("lnorm", "bs", "gamma", "geom", "ls",
  "mzsm", "nbinom", "pareto", "poilog", "power", "volkov", "powbend",
  "weibull"), sad_coef = list(cv_abund = 1), fix_s_sim = FALSE)
}
\arguments{
\item{s_pool}{Number of species in the pool (integer)}

\item{n_sim}{Number of individuals in the simulated community (integer)}

\item{sad_type}{Root name of community sad distribution - e.g., lnorm for
  the lognormal distribution (\code{\link[stats]{rlnorm}});
  geom for the geometric distribution (\code{\link[stats]{rgeom}}),
  or ls for the log-series distribution \code{\link[sads]{rls}}).

  See \code{\link[sads]{rsad}} for all options. (character)}

\item{sad_coef}{List with named arguments to be passed to the probability
  function defined by the argument \code{sad_type}. The parameter names
  are given in the respective distribution functions. See e.g.
  \code{\link[sads]{rls}} for the parameters of the log-series, or
  \code{\link[stats]{rlnorm}} for the parameters of the log-normal
  distribution.

  In \code{mobsim} the log-normal distribution can alternatively be
  parameterized by the coefficient of variation of the abundances
  \code{cv_abund}, which is the standard deviation of abundances divided by
  the mean abundance (no. of individuals / no. of species). \code{cv_abund}
  is thus negatively correlated with the evenness of the species abundance
  distribution.}

\item{fix_s_sim}{Should the simulation constrain the number of
species in the simulated local community? (logical)}
}
\value{
Object of class \code{sad}, which contains a named integer vector
 with species abundances
}
\description{
Simulate species abundance with user-defined number
of species in the pool \code{(s_pool)}, and number of individuals
\code{(n_sim)} in the simulated community.
}
\details{
The function \code{sim_sad} provides a wrapper around the function
  \code{\link[sads]{rsad}} from the R package \code{\link{sads}}. However, in
  contrast to \code{\link[sads]{rsad}}, the function \code{sim_sad} allows to
  define the number of individuals in the simulated community. This is
  implemented by converting the abundance distribution simulated with
  \code{\link[sads]{rsad}} into a relative abundance distribution. In a second
  step the required no. of individuals \code{(n_sim)} is sampled from this
  distribution (with replacement).

  Please note that this might effect the interpretation of the parameters of
  the underlying statistical distribution, e.g. the mean abundance will always
  be \code{n_sim/n_pool} irrespective of the settings of \code{sad_coef}.

When \code{fix_s_sim = FALSE} the species number in the local
  community might deviate from \code{s_pool} due to stochastic sampling. When
  \code{fix_s_sim = TRUE} the local number of species will equal
  \code{s_pool}, but this constraint can result in systematic biases from the
  theoretical distribution parameters. Generally, with \code{fix_s_sim = TRUE}
  additional very rare species will be added to the community, while the abundance
  of the most common ones is reduced to keep the defined number of individuals.
}
\examples{
abund1 <- sim_sad(s_pool = 100, n_sim = 10000, sad_type = "lnorm",
 sad_coef = list("cv_abund" = 1))
plot(abund1, method = "octave")
plot(abund1, method = "rank")

}
\author{
Felix May
}
