% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{sim_thomas_community}
\alias{sim_thomas_community}
\title{Simulate community with clumped spatial positions.}
\usage{
sim_thomas_community(s_pool, n_sim, sad_type = c("lnorm", "bs", "gamma",
  "geom", "ls", "mzsm", "nbinom", "pareto", "poilog", "power", "volkov",
  "powbend", "weibull"), sad_coef = list(cv_abund = 1), fix_s_sim = FALSE,
  sigma = 0.02, cluster_points = NA, mother_points = NA, xrange = c(0,
  1), yrange = c(0, 1))
}
\arguments{
\item{s_pool}{Number of species in the pool (integer)}

\item{n_sim}{Number of individuals in the simulated community (integer)}

\item{sad_type}{Root name of community sad distribution - e.g., lnorm for
  the lognormal distribution (\code{\link[stats]{rlnorm}});
  geom for the geometric distribution \code{\link[stats]{rgeom}}),
  or ls for the log-series distribution \code{\link[sads]{rls}}).

  See \code{\link[sads]{rsad}} for all options. (character)}

\item{sad_coef}{List with named arguments to be passed to the probability
  function defined by the argument \code{sad_type}. The parameter names
  are given in the respective distribution functions. See e.g.
  \code{\link[sads]{rls}} for the parameters of the log-series, or
  \code{\link[stats]{rlnorm}} for the parameters of the log-normal
  distribution.

  In \code{mobsim} the log-normal distribution can alternatively be
  parameterized by the coefficient of variation of the abundances
  \code{cv_abund}, which is the standard deviation of abundances divided by
  the mean abundance (no. of individuals / no. of species). \code{cv_abund}
  is this negatively correlated with community evenness.}

\item{fix_s_sim}{Should the simulation constrain the number of
species in the simulated local community? (logical)}

\item{sigma}{Mean displacement (along each coordinate axes) of a point from
its mother point (= cluster centre).}

\item{cluster_points}{Mean number of points per cluster.}

\item{mother_points}{Number of mother points (= cluster centres).}

\item{xrange}{Extent of the community in x-direction (numeric vector of length 2)}

\item{yrange}{Extent of the community in y-direction (numeric vector of length 2)}
}
\value{
A community object as defined by \code{\link{community}}
}
\description{
This function simulates a community with a certain abundance distribution and
and clumped, i.e. aggregated, spatial coordinates. This function consecutively calls
\code{\link{sim_sad}} and \code{\link{sim_thomas_coords}}
}
\details{
See the documentations of \code{\link{sim_sad}} and
 \code{\link{sim_thomas_coords}} for details.
}
\examples{
com1 <- sim_thomas_community(s_pool = 20, n_sim = 500, sad_type = "lnorm",
                             sad_coef = list("meanlog" = 2, "sdlog" = 1),
                             sigma = 0.01)
plot(com1)

}
\author{
Felix May
}
