% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{sim_thomas_community}
\alias{sim_thomas_community}
\title{Simulate community with clumped spatial positions.}
\usage{
sim_thomas_community(s_pool, n_sim, cv_abund = 1, fix_s_local = F,
  sigma = 0.02, cluster_points = NA, mother_points = NA, xrange = c(0,
  1), yrange = c(0, 1))
}
\arguments{
\item{s_pool}{integer - number of species in the pool}

\item{n_sim}{integer - number of individuals in the simulate local community}

\item{cv_abund}{numeric - coefficient of variation ( = sd/mean) of relative
abundances. The higher \code{cv_abund}, the lower the evenness of the
simulated community. This means with increasing \code{cv_abund} there are more
rare and more dominant species.}

\item{fix_s_local}{logical - should the simulation constrain the number of
species in the local community? This can result in deviations from mean and
sd of local abundances from the theoretical distributions}

\item{sigma}{numeric vector of length = 1 or length = 2. Mean displacement
(along each coordinate axes) of a point from its mother point (=cluster centre).}

\item{cluster_points}{numeric - mean number of points per cluster.}

\item{mother_points}{numeric - number of mother points (= cluster centres).}

\item{xrange}{numeric vector of length 2 - extent of the community in x-direction}

\item{yrange}{numeric vector of length 2 - extent of the community in y-direction}
}
\value{
A community object as defined by \code{\link{community}}.
}
\description{
This function simulates a community with a certain abundance distribution and
and clumped, i.e. aggregated, spatial coordinates. This function consecutively calls
\code{\link{sim_sad}} and \code{\link{sim_thomas_coords}}
}
\examples{
com1 <- sim_thomas_community(s_pool = 20, n_sim = 500, cv_abund = 1,
                             sigma = 0.01)
plot(com1)

}
\author{
Felix May
}
\seealso{
\code{sim_sad}; \code{sim_thomas_coords}
}

