% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{summary.community}
\alias{summary.community}
\title{Print summary of spatial community object}
\usage{
\method{summary}{community}(object, digits = 2, ...)
}
\arguments{
\item{object}{Community object of class \code{\link{community}}}

\item{digits}{Integer. Number of digits to print}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\value{
This function is called for its side effects and has no return value.
}
\description{
Print summary of spatial community object
}
