% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Community.R
\name{summary.community}
\alias{summary.community}
\title{Print summary of spatial community object}
\usage{
\method{summary}{community}(object, ...)
}
\arguments{
\item{object}{Community object of class \code{\link{community}}}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
Print summary of spatial community object
}
